chore: Move logic to hook
