+++ src/hooks/index.ts
@@ -16,3 +16,4 @@ export * from './useStyles';
 export * from './useTheme';
 export * from './useUser';
 export * from './useWearables';
+export * from './useEnvironmentToggle';
